requirements.txt
pytest
selenium
jsonpickle
allure-pytest
